requirements.txt
kiteconnect==4.2.0
numpy==1.24.3
scikit-learn==1.3.0
python-telegram-bot==20.5
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
redis==4.6.0
flask==2.3.2
requests==2.31.0
python-dotenv==1.0.0
pytz==2023.3
loguru==0.7.0
ta==0.10.2
schedule==1.2.0
asyncio-mqtt==0.16.1
pandas>=2.0.0
websockets>=10.0
matplotlib>=3.5.0
chart.js>=3.9.1